NCI-Agency/anet#1705: Refactor PostgreSQL report search so batching works
Also, since there's no longer a need to support SQLite, we can pull up a lot more search code since PostgreSQL and MS SQL have much more in common.
