+++ src/main/java/mil/dds/anet/search/AbstractTaskSearcher.java
@@ -58,9 +58,33 @@ public abstract class AbstractTaskSearcher extends AbstractSearcher<Task, TaskSe
 
   protected abstract void addTextQuery(TaskSearchQuery query);
 
-  protected abstract void addResponsibleOrgUuidQuery(TaskSearchQuery query);
+  protected void addResponsibleOrgUuidQuery(TaskSearchQuery query) {
+    if (Boolean.TRUE.equals(query.getIncludeChildrenOrgs())) {
+      qb.addWithClause("parent_orgs(uuid) AS ("
+          + " SELECT uuid FROM organizations WHERE uuid = :orgUuid UNION ALL"
+          + " SELECT o.uuid FROM parent_orgs po, organizations o WHERE o.\"parentOrgUuid\" = po.uuid"
+          + ")");
+      qb.addWhereClause("tasks.\"organizationUuid\" IN (SELECT uuid FROM parent_orgs)");
+      qb.addSqlArg("orgUuid", query.getResponsibleOrgUuid());
+    } else {
+      qb.addEqualsClause("orgUuid", "tasks.\"organizationUuid\"", query.getResponsibleOrgUuid());
+    }
+  }
 
-  protected abstract void addCustomFieldRef1UuidQuery(TaskSearchQuery query);
+  protected void addCustomFieldRef1UuidQuery(TaskSearchQuery query) {
+    if (Boolean.TRUE.equals(query.getCustomFieldRef1Recursively())) {
+      qb.addWithClause("parent_tasks(uuid) AS ("
+          + " SELECT uuid FROM tasks WHERE uuid = :customFieldRef1Uuid UNION ALL"
+          + " SELECT t.uuid FROM parent_tasks pt, tasks t WHERE t.\"customFieldRef1Uuid\" = pt.uuid AND t.uuid != :customFieldRef1Uuid"
+          + ")");
+      qb.addWhereClause("(tasks.\"customFieldRef1Uuid\" IN (SELECT uuid FROM parent_tasks)"
+          + " OR tasks.uuid = :customFieldRef1Uuid)");
+      qb.addSqlArg("customFieldRef1Uuid", query.getCustomFieldRef1Uuid());
+    } else {
+      qb.addEqualsClause("customFieldRef1Uuid", "tasks.\"customFieldRef1Uuid\"",
+          query.getCustomFieldRef1Uuid());
+    }
+  }
 
   protected void addOrderByClauses(AbstractSearchQueryBuilder<?, ?> qb, TaskSearchQuery query) {
     switch (query.getSortBy()) {
