+++ src/main/java/mil/dds/anet/search/mssql/MssqlOrganizationSearcher.java
@@ -3,7 +3,6 @@ package mil.dds.anet.search.mssql;
 import mil.dds.anet.beans.Organization;
 import mil.dds.anet.beans.search.ISearchQuery.SortOrder;
 import mil.dds.anet.beans.search.OrganizationSearchQuery;
-import mil.dds.anet.database.OrganizationDao;
 import mil.dds.anet.search.AbstractOrganizationSearcher;
 import mil.dds.anet.search.AbstractSearchQueryBuilder;
 
@@ -12,13 +11,6 @@ public class MssqlOrganizationSearcher extends AbstractOrganizationSearcher {
   public MssqlOrganizationSearcher() {
     super(new MssqlSearchQueryBuilder<Organization, OrganizationSearchQuery>(
         "MssqlOrganizationSearch"));
-  }
-
-  @Override
-  protected void buildQuery(OrganizationSearchQuery query) {
-    qb.addSelectClause(OrganizationDao.ORGANIZATION_FIELDS);
-    super.buildQuery(query);
-    qb.addTotalCount();
   }
 
   @Override
@@ -39,21 +31,6 @@ public class MssqlOrganizationSearcher extends AbstractOrganizationSearcher {
     final String text = query.getText();
     qb.addSqlArg("containsQuery", qb.getFullTextQuery(text));
     qb.addSqlArg("likeQuery", qb.getLikeQuery(text));
-  }
-
-  @Override
-  protected void addParentOrgUuidQuery(OrganizationSearchQuery query) {
-    if (Boolean.TRUE.equals(query.getParentOrgRecursively())) {
-      qb.addWithClause("parent_orgs(uuid) AS ("
-          + " SELECT uuid FROM organizations WHERE uuid = :parentOrgUuid UNION ALL"
-          + " SELECT o.uuid FROM parent_orgs po, organizations o WHERE o.parentOrgUuid = po.uuid AND o.uuid != :parentOrgUuid"
-          + ")");
-      qb.addWhereClause("(organizations.parentOrgUuid IN (SELECT uuid FROM parent_orgs)"
-          + " OR organizations.uuid = :parentOrgUuid)");
-    } else {
-      qb.addWhereClause("organizations.parentOrgUuid = :parentOrgUuid");
-    }
-    qb.addSqlArg("parentOrgUuid", query.getParentOrgUuid());
   }
 
   @Override
