+++ src/main/java/mil/dds/anet/search/mssql/MssqlReportSearcher.java
@@ -1,15 +1,11 @@
 package mil.dds.anet.search.mssql;
 
-import mil.dds.anet.beans.Organization;
 import mil.dds.anet.beans.Report;
 import mil.dds.anet.beans.lists.AnetBeanList;
 import mil.dds.anet.beans.search.ISearchQuery.SortOrder;
 import mil.dds.anet.beans.search.ReportSearchQuery;
-import mil.dds.anet.database.ReportDao;
-import mil.dds.anet.database.mappers.ReportMapper;
 import mil.dds.anet.search.AbstractReportSearcher;
 import mil.dds.anet.search.AbstractSearchQueryBuilder;
-import ru.vyarus.guicey.jdbi3.tx.InTransaction;
 
 public class MssqlReportSearcher extends AbstractReportSearcher {
 
@@ -21,26 +17,8 @@ public class MssqlReportSearcher extends AbstractReportSearcher {
   }
 
   @Override
-  protected void buildQuery(ReportSearchQuery query) {
-    qb.addSelectClause("DISTINCT " + ReportDao.REPORT_FIELDS);
-    super.buildQuery(query);
-  }
-
-  @InTransaction
-  @Override
   public AnetBeanList<Report> runSearch(ReportSearchQuery query) {
-    buildQuery(query);
-    outerQb.addSelectClause("*");
-    outerQb.addTotalCount();
-    outerQb.addFromClause("( " + qb.build() + " ) l");
-    outerQb.addSqlArgs(qb.getSqlArgs());
-    outerQb.addListArgs(qb.getListArgs());
-    final AnetBeanList<Report> result =
-        outerQb.buildAndRun(getDbHandle(), query, new ReportMapper());
-    for (final Report report : result.getList()) {
-      report.setUser(query.getUser());
-    }
-    return result;
+    return runSearch(outerQb, query);
   }
 
   @Override
@@ -81,82 +59,31 @@ public class MssqlReportSearcher extends AbstractReportSearcher {
 
   @Override
   protected void addOrgUuidQuery(ReportSearchQuery query) {
-    if (!query.getIncludeOrgChildren()) {
-      qb.addWhereClause(
-          "(reports.advisorOrganizationUuid = :orgUuid OR reports.principalOrganizationUuid = :orgUuid)");
-    } else {
-      outerQb.addWithClause("parent_orgs(uuid) AS ("
-          + " SELECT uuid FROM organizations WHERE uuid = :orgUuid UNION ALL"
-          + " SELECT o.uuid FROM parent_orgs po, organizations o WHERE o.parentOrgUuid = po.uuid"
-          + ")");
-      qb.addWhereClause("(reports.advisorOrganizationUuid IN (SELECT uuid FROM parent_orgs)"
-          + " OR reports.principalOrganizationUuid IN (SELECT uuid FROM parent_orgs))");
-    }
-    qb.addSqlArg("orgUuid", query.getOrgUuid());
+    addOrgUuidQuery(outerQb, query);
   }
 
   @Override
   protected void addAdvisorOrgUuidQuery(ReportSearchQuery query) {
-    if (Organization.DUMMY_ORG_UUID.equals(query.getAdvisorOrgUuid())) {
-      qb.addWhereClause("reports.advisorOrganizationUuid IS NULL");
-    } else if (!query.getIncludeAdvisorOrgChildren()) {
-      qb.addEqualsClause("advisorOrganizationUuid", "reports.advisorOrganizationUuid",
-          query.getAdvisorOrgUuid());
-    } else {
-      outerQb.addWithClause("advisor_parent_orgs(uuid) AS ("
-          + " SELECT uuid FROM organizations WHERE uuid = :advisorOrgUuid UNION ALL"
-          + " SELECT o.uuid FROM advisor_parent_orgs po, organizations o WHERE o.parentOrgUuid = po.uuid"
-          + ")");
-      qb.addWhereClause(
-          "reports.advisorOrganizationUuid IN (SELECT uuid FROM advisor_parent_orgs)");
-      qb.addSqlArg("advisorOrgUuid", query.getAdvisorOrgUuid());
-    }
+    addAdvisorOrgUuidQuery(outerQb, query);
   }
 
   @Override
   protected void addPrincipalOrgUuidQuery(ReportSearchQuery query) {
-    if (Organization.DUMMY_ORG_UUID.equals(query.getPrincipalOrgUuid())) {
-      qb.addWhereClause("reports.principalOrganizationUuid IS NULL");
-    } else if (!query.getIncludePrincipalOrgChildren()) {
-      qb.addEqualsClause("principalOrganizationUuid", "reports.principalOrganizationUuid",
-          query.getPrincipalOrgUuid());
-    } else {
-      outerQb.addWithClause("principal_parent_orgs(uuid) AS ("
-          + " SELECT uuid FROM organizations WHERE uuid = :principalOrgUuid UNION ALL"
-          + " SELECT o.uuid FROM principal_parent_orgs po, organizations o WHERE o.parentOrgUuid = po.uuid"
-          + ")");
-      qb.addWhereClause(
-          "reports.principalOrganizationUuid IN (SELECT uuid FROM principal_parent_orgs)");
-      qb.addSqlArg("principalOrgUuid", query.getPrincipalOrgUuid());
-    }
+    addPrincipalOrgUuidQuery(outerQb, query);
   }
 
   @Override
   protected void addOrderByClauses(AbstractSearchQueryBuilder<?, ?> qb, ReportSearchQuery query) {
-    if (query.isTextPresent() && !query.isSortByPresent()) {
-      // We're doing a full-text search without an explicit sort order,
-      // so sort first on the search pseudo-rank.
-      outerQb.addAllOrderByClauses(getOrderBy(SortOrder.DESC, null, "search_rank"));
-    }
+    if (qb == outerQb) {
+      // ordering must be on the outer query
+      if (query.isTextPresent() && !query.isSortByPresent()) {
+        // We're doing a full-text search without an explicit sort order,
+        // so sort first on the search pseudo-rank.
+        qb.addAllOrderByClauses(getOrderBy(SortOrder.DESC, null, "search_rank"));
+      }
 
-    // Beware of the sort field names, they have to match what's in the selected fields!
-    switch (query.getSortBy()) {
-      case CREATED_AT:
-        outerQb.addAllOrderByClauses(getOrderBy(query.getSortOrder(), null, "reports_createdAt"));
-        break;
-      case RELEASED_AT:
-        outerQb.addAllOrderByClauses(getOrderBy(query.getSortOrder(), null, "reports_releasedAt"));
-        break;
-      case UPDATED_AT:
-        outerQb.addAllOrderByClauses(getOrderBy(query.getSortOrder(), null, "reports_updatedAt"));
-        break;
-      case ENGAGEMENT_DATE:
-      default:
-        outerQb
-            .addAllOrderByClauses(getOrderBy(query.getSortOrder(), null, "reports_engagementDate"));
-        break;
+      super.addOrderByClauses(qb, query);
     }
-    outerQb.addAllOrderByClauses(getOrderBy(SortOrder.ASC, null, "reports_uuid"));
   }
 
 }
